ERROR WILL BE GONE 2 ELECTRIC BOOGALOO

+++ games/flappyBird/main_flappy_bird.py
@@ -122,7 +122,7 @@ def main():
         if not bird_player.dead:
             if add_pole_timer >= add_pole_cooldown:
                 add_pole_timer = 0
-                y = random.randrange(-pole_height + screen_height/10, 0)
+                y = random.randrange(-pole_height + int(screen_height/10), 0)
                 entities.append(pole.Pole((screen_width, y), 250, False))
             else:
                 add_pole_timer += 1
